+++ tests/testthat/test-marginal_tidiers.R
@@ -164,7 +164,7 @@ test_that("tidy_marginal_predictions()", {
     res <- tidy_plus_plus(
       mod,
       tidy_fun = tidy_marginal_predictions,
-      newdata = "marginalmeans"
+      newdata = "balanced"
     )
   )
   expect_equal(
@@ -232,7 +232,7 @@ test_that("tidy_avg_slopes()", {
     res <- tidy_plus_plus(
       mod,
       tidy_fun = tidy_avg_slopes,
-      newdata = "marginalmeans"
+      newdata = "balanced"
     )
   )
   expect_equal(
@@ -307,22 +307,12 @@ test_that("tidy_marginal_contrasts()", {
     res <- tidy_plus_plus(
       mod,
       tidy_fun = tidy_marginal_contrasts,
-      newdata = "marginalmeans"
+      newdata = "balanced"
     )
   )
   expect_equal(
     attr(res, "coefficients_label"),
     "Marginal Contrasts at Marginal Means"
-  )
-})
-
-test_that("tidy_marginal_means()", {
-  skip_on_cran()
-  skip_if_not_installed("marginaleffects")
-
-  mod <- lm(Petal.Length ~ Petal.Width * Species + Sepal.Length, data = iris)
-  lifecycle::expect_deprecated(
-    tidy_marginal_means(mod)
   )
 })
 
@@ -375,7 +365,7 @@ test_that("tidy_avg_comparisons()", {
     res <- tidy_plus_plus(
       mod,
       tidy_fun = tidy_avg_comparisons,
-      newdata = "marginalmeans"
+      newdata = "balanced"
     )
   )
   expect_equal(